fix: rename vendored breeze-tts 'models' package to 'breeze_models' to avoid global editable-install name collision
