+++ ext/py/breeze-tts/infer.py
@@ -16,8 +16,8 @@ from breeze_infer.runtime import (
     update_generation_config_for_breeze,
 )
 from breeze_infer.templates import get_template, prepare_inputs
-from models.fast_streaming import FastBreezeStreamingRuntime, FastStreamingConfig
-from models.warmup_profile import load_warmup_profile
+from breeze_models.fast_streaming import FastBreezeStreamingRuntime, FastStreamingConfig
+from breeze_models.warmup_profile import load_warmup_profile
 
 REPO_ROOT = Path(__file__).resolve().parent
 FAST_CONFIG = REPO_ROOT / "configs" / "fast.json"
