+++ ext/py/breeze-tts/setup.py
@@ -2,7 +2,9 @@
 # setup.py/pyproject.toml, so it is not pip-installable as-is. This file was authored
 # for this vendored copy (ext/py/breeze-tts) so it can be installed in editable mode
 # the same way ./ext/py/demucs already is in this repo's requirements.txt - it is not
-# copied from upstream.
+# copied from upstream. Upstream's top-level `models` package was renamed to
+# `breeze_models` here (and all internal imports updated to match) since `models`
+# is too generic a name to install globally editable without risking collisions.
 from setuptools import setup, find_packages
 
 setup(
