+++ ext/py/breeze-tts/tests/test_fast_streaming.py
@@ -6,7 +6,7 @@ import pytest
 import torch
 
 from breeze_infer.templates import get_template
-from models.fast_streaming import (
+from breeze_models.fast_streaming import (
     FastBreezeStreamingRuntime,
     FastStreamingConfig,
     _get_dtype,
