+++ ext/py/breeze-tts/tests/test_logits_process.py
@@ -2,8 +2,8 @@ from __future__ import annotations
 
 import torch
 
-from models.cudagraph.sampling import sample_logits
-from models.logits_process import (
+from breeze_models.cudagraph.sampling import sample_logits
+from breeze_models.logits_process import (
     GeneratedTokenRepetitionPenaltyLogitsProcessor,
     mask_invalid_codec_token_logits,
 )
